+++ src/renderers/passes/index.ts
@@ -4,3 +4,4 @@ export * from "./MainRenderPass.js";
 export * from "./PostProcessPass.js";
 export * from "./CascadedShadowPassGPU.js";
 export * from "./SpotShadowPassGPU.js";
+export * from "./DepthPrePassGPU.js";